services/products/requirements.txt
Flask
Flask-Caching
redis
pandas
Flask-CORS
gunicorn
psycopg2-binary
python-dotenv
Werkzeug==2.3.7
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0


